requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.45
starlette==0.50.0
style==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
update==0.0.1
urllib3==2.6.2
uvicorn==0.40.0
pytest==8.3.4
httpx==0.28.1
